Moved Connections.cs, Stations.cs and StationBoardRoot.cs into Models/*
