+++ src/SwissTransport/Models/Connections.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using Newtonsoft.Json;
 
-namespace SwissTransport
+namespace SwissTransport.Models
 {
     public class Connections
     {
