+++ src/SwissTransport/SwissTransport.csproj
@@ -44,9 +44,9 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Connections.cs" />
-    <Compile Include="StationBoardRoot.cs" />
-    <Compile Include="Stations.cs" />
+    <Compile Include="Models\Connections.cs" />
+    <Compile Include="Models\StationBoardRoot.cs" />
+    <Compile Include="Models\Stations.cs" />
     <Compile Include="ITransport.cs" />
     <Compile Include="Transport.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
