+++ src/SwissTransport/Transport.cs
@@ -1,6 +1,7 @@
 ﻿using System.IO;
 using System.Net;
 using Newtonsoft.Json;
+using SwissTransport.Models;
 
 namespace SwissTransport
 {
